+++ modules/gui/debug_edit_party.py
@@ -1,17 +1,11 @@
-import random
-import struct
 import time
 import tkinter
 from tkinter import Tk, Toplevel, ttk
-from typing import Literal
-
-import numpy
 
 from modules.context import context
-from modules.game import encode_string
-from modules.items import get_item_by_index, Item, get_item_by_name
+from modules.debug_utilities import debug_create_pokemon, debug_write_party
+from modules.items import get_item_by_index, get_item_by_name
 from modules.memory import write_symbol
-from modules.player import get_player
 from modules.pokemon import (
     get_party,
     Pokemon,
@@ -20,13 +14,10 @@ from modules.pokemon import (
     get_move_by_index,
     StatsValues,
     Species,
-    POKEMON_DATA_SUBSTRUCTS_ORDER,
-    Nature,
     StatusCondition,
     get_move_by_name,
     get_species_by_name,
 )
-from modules.roms import ROMLanguage
 
 status_name_map = {
     "Healthy": StatusCondition.Healthy,
@@ -93,12 +84,7 @@ class PartyEditMenu:
             if pokemon is not None:
                 new_party.append(pokemon)
 
-        if len(new_party) > 0:
-            write_symbol("gPlayerPartyCount", len(new_party).to_bytes(1, byteorder="little"))
-            new_party_data = b""
-            for pokemon in new_party:
-                new_party_data += pokemon.data
-            write_symbol("gPlayerParty", new_party_data.ljust(600, b"\x00"))
+        debug_write_party(new_party)
 
         self.close_window()
 
@@ -197,7 +183,7 @@ class PokemonEditFrame:
         while len(moves) < 4:
             moves.append({"id": 0, "remaining_pp": 0, "pp_ups": 0})
 
-        return _create_pokemon(
+        return debug_create_pokemon(
             self._pokemon,
             is_shiny=self._is_shiny_var.get(),
             is_egg=self._is_egg_var.get(),
@@ -655,215 +641,6 @@ class PokemonEditFrame:
             exp_window.update_idletasks()
             exp_window.update()
             time.sleep(1 / 60)
-
-
-def _create_pokemon(
-    original_pokemon: Pokemon,
-    is_egg: bool,
-    is_shiny: bool,
-    gender: Literal["male", "female"] | None,
-    species: Species,
-    nickname: str,
-    level: int,
-    held_item: Item | None,
-    has_second_ability: bool,
-    nature: Nature,
-    experience: int,
-    friendship: int,
-    moves: list[dict],
-    ivs: StatsValues,
-    evs: StatsValues,
-    current_hp: int,
-    status_condition: StatusCondition,
-) -> Pokemon:
-    iv_egg_ability = (
-        ivs.hp
-        | (ivs.attack << 5)
-        | (ivs.defence << 10)
-        | (ivs.speed << 15)
-        | (ivs.special_attack << 20)
-        | (ivs.special_defence << 25)
-    )
-    if has_second_ability:
-        iv_egg_ability |= 1 << 31
-    if is_egg:
-        iv_egg_ability |= 1 << 30
-
-    pp_bonuses = (
-        (moves[0]["pp_ups"] << 6) | (moves[1]["pp_ups"] << 4) | (moves[2]["pp_ups"] << 2) | (moves[3]["pp_ups"] << 0)
-    )
-
-    if original_pokemon.is_empty:
-        player = get_player()
-
-        match context.rom.language:
-            case ROMLanguage.Japanese:
-                language = 1
-            case ROMLanguage.French:
-                language = 3
-            case ROMLanguage.Italian:
-                language = 4
-            case ROMLanguage.German:
-                language = 5
-            case ROMLanguage.Spanish:
-                language = 7
-            case _:
-                language = 2
-
-        original_pokemon = Pokemon(
-            # Personality Value, gets generated later
-            (b"\x00" * 4)
-            + player.trainer_id.to_bytes(2, byteorder="little")
-            + player.secret_id.to_bytes(2, byteorder="little")
-            # Nickname, gets generated later
-            + (b"\x00" * 10)
-            + language.to_bytes(1)
-            + (b"\x06" if is_egg else b"\x02")
-            + encode_string(player.name).ljust(7, b"\xFF")
-            + (b"\x00" * 73)
-        )
-
-        if context.rom.is_sapphire:
-            game_number = 1
-        elif context.rom.is_ruby:
-            game_number = 2
-        elif context.rom.is_emerald:
-            game_number = 3
-        elif context.rom.is_fr:
-            game_number = 4
-        elif context.rom.is_lg:
-            game_number = 5
-        else:
-            game_number = 15
-        decrypted_data = (
-            # First 3 blocks don't matter.
-            (b"\x00" * (12 * 3 + 32))
-            # Pokérus Status
-            + b"\x00"
-            # Met Location (0x88 and 0x39 are the Kanto and Hoenn Safari Zones)
-            + (b"\x88" if context.rom.is_frlg else b"\x39")
-            + (
-                # OT gender
-                bool(player.gender == "female") << 15
-                |
-                # Ball (12 = Premier Ball)
-                (12 << 11)
-                |
-                # Game of Origin
-                (game_number << 7)
-                |
-                # Met at Level
-                level
-            ).to_bytes(2, "little")
-            + (b"\x00" * 4)
-            + (0 if species.name not in ("Mew", "Deoxys") else 1 << 31).to_bytes(4, "little")
-            + (b"\x00" * 20)
-        )
-    else:
-        decrypted_data = original_pokemon._decrypted_data
-
-    if is_egg and friendship > species.egg_cycles:
-        friendship = species.egg_cycles
-
-    data_to_encrypt = (
-        species.index.to_bytes(2, byteorder="little")
-        + (held_item.index if held_item is not None else 0).to_bytes(2, byteorder="little")
-        + experience.to_bytes(4, byteorder="little")
-        + pp_bonuses.to_bytes(1)
-        + friendship.to_bytes(1)
-        + decrypted_data[42:44]
-        + moves[0]["id"].to_bytes(2, byteorder="little")
-        + moves[1]["id"].to_bytes(2, byteorder="little")
-        + moves[2]["id"].to_bytes(2, byteorder="little")
-        + moves[3]["id"].to_bytes(2, byteorder="little")
-        + moves[0]["remaining_pp"].to_bytes(1)
-        + moves[1]["remaining_pp"].to_bytes(1)
-        + moves[2]["remaining_pp"].to_bytes(1)
-        + moves[3]["remaining_pp"].to_bytes(1)
-        + evs.hp.to_bytes(1)
-        + evs.attack.to_bytes(1)
-        + evs.defence.to_bytes(1)
-        + evs.speed.to_bytes(1)
-        + evs.special_attack.to_bytes(1)
-        + evs.special_defence.to_bytes(1)
-        + decrypted_data[62:72]
-        + iv_egg_ability.to_bytes(4, byteorder="little")
-        + decrypted_data[76:80]
-    )
-
-    if nickname != "" and nickname != species.name.upper():
-        encoded_nickname = encode_string(nickname)
-    else:
-        encoded_nickname = encode_string(species.name.upper())
-
-    stats = StatsValues.calculate(species, ivs, evs, nature, level)
-
-    def personality_value_matches_criteria(pv: int) -> bool:
-        if pv % 25 != nature.index:
-            return False
-
-        if 0 < species.gender_ratio < 254:
-            is_male = pv & 0xFF >= species.gender_ratio
-            if is_male and gender != "male":
-                return False
-            if not is_male and gender == "male":
-                return False
-
-        shiny_value = (
-            original_pokemon.original_trainer.id
-            ^ original_pokemon.original_trainer.secret_id
-            ^ (pv >> 16)
-            ^ (pv & 0xFFFF)
-        )
-        if is_shiny and shiny_value > 7:
-            return False
-        if not is_shiny and shiny_value <= 7:
-            return False
-
-        return True
-
-    personality_value = original_pokemon.personality_value
-    n = 0
-    while not personality_value_matches_criteria(personality_value):
-        if n > 10_000_000:
-            raise RuntimeError("Could not find a suitable PV in time.")
-        n += 1
-        personality_value = random.randint(0, 2**32)
-
-    data = (
-        personality_value.to_bytes(length=4, byteorder="little")
-        + original_pokemon.data[4:8]
-        + encoded_nickname.ljust(10, b"\xFF")
-        + original_pokemon.data[18:19]
-        + (b"\x06" if is_egg else b"\x02")
-        + original_pokemon.data[20:28]
-        + (sum(struct.unpack("<24H", data_to_encrypt)) & 0xFFFF).to_bytes(length=2, byteorder="little")
-        + original_pokemon.data[30:32]
-        + data_to_encrypt
-        + status_condition.to_bitfield().to_bytes(length=1, byteorder="little")
-        + original_pokemon.data[81:84]
-        + level.to_bytes(length=1, byteorder="little")
-        + original_pokemon.data[85:86]
-        + current_hp.to_bytes(length=2, byteorder="little")
-        + stats.hp.to_bytes(length=2, byteorder="little")
-        + stats.attack.to_bytes(length=2, byteorder="little")
-        + stats.defence.to_bytes(length=2, byteorder="little")
-        + stats.speed.to_bytes(length=2, byteorder="little")
-        + stats.special_attack.to_bytes(length=2, byteorder="little")
-        + stats.special_defence.to_bytes(length=2, byteorder="little")
-    )
-
-    u32le = numpy.dtype("<u4")
-    personality_value_bytes = numpy.frombuffer(data, count=1, dtype=u32le)
-    original_trainer_id = numpy.frombuffer(data, count=1, offset=4, dtype=u32le)
-    key = numpy.repeat(personality_value_bytes ^ original_trainer_id, 3)
-    order = POKEMON_DATA_SUBSTRUCTS_ORDER[personality_value % 24]
-
-    encrypted_data = numpy.concatenate(
-        [numpy.frombuffer(data, count=3, offset=32 + (order.index(i) * 12), dtype=u32le) ^ key for i in range(4)]
-    )
-
-    return Pokemon(data[:32] + encrypted_data.tobytes() + data[80:100])
 
 
 def run_edit_party_screen():
